Added expert option to disable snapshots when on battery: https://bugs.launchpad.net/backintime/+bug/388178
This involves using the on_ac_power command to detect if the power status
is available.
If it is not available then the expert option is disabled.
If it is available the user has the option.
When the backintime command is run by cron it checks the power status
and defers the snapshot if the system is on battery power.

Changes include two new functions to the tools module for power status
checking, changes to the two settings windows for Gnome & KDE, changes
to the config system to save the option and finally the check in the
snapshot module.
For debian packages the powermgmt-base package is recommended since it
contains the on_ac_power command.