harden serviceHelper by not allowing && and only well-known commands
the addRule() dbus method is very broad in allowing arbitrary commands
to be added to udev. The serviceHelper makes sure that those commands
are always safely executed via 'su' as the user that registered the
rule. So no privilege escalation should be possible.

Still the possibility to execute just any command in the context of the
user might prove to be security sensitive in combination with future
weaknesses or other weaknesses not yet fully understood.

This commit tightens the range of allowed commands by only allowing well
known commands. A combination of nice, ionice and backintime itself
using full paths is allowed an nothing more.

Actually it'd be better to change the addRule() interface in a way that
only the nice/ionice parameters (if any) and the parameters to
backintime should be passed instead of an arbitrary command line. I
didn't want to introduce such a big change for this hardening measure,
however.