Resolve notify interface dynamically
If backintime runs longer than the notification daemon (or starts before
it), it may miss to detect it properly.

This change will add a small overhead by resolving the notify interface
for each notification, however it will make backintime send
notifications when possible (again) and will still be faster than
calling notify-send (which had to do the same).