feat: Extend diagnostic information's
1. Python executable path.
2. Rsync environment variables.
3. Modified username replacement to replace users homepath and not only
   its username.

About 1: Related to #1316 where the path to systems python interpreter
was "hacked" via "pyenv". This info will catch situations like that.

About 2: Related to #1247. Rsync 3.2.4 do use a different way of
argument protection. The situation is complex and the environment
variables RSYNC_OLD_ARGS and RSYNC_PROTECT_ARGS to influence that.

About 3: The replacement is done on the whole JSON-like string and not
on specific elements. To prevent replacement of technical informations
(e.g. about "rsync") when a user do use a technical login name (e.g.
"rsync") now we search for the users homepath and not only is username.