enh: New test about remove snapshots via SSH.
Introducing a new unit (or better integration) test about removing a
snapshot via SSH because there wasn't something like this before. I need
this test because of the _rsync incompatibility problem_ (#1247). It is
intended to fail with a newer rsync version and to pass with an old one.
I checked it with _Debian testing_ (_rsync 3.2.6_) and _Debian 11 stable
"bullseye"_ (_rsync 3.2.3_).
The target of the test is the method `snapshots.Snapsthots.remove()`. As
I explain there in the doc string the removing method seems a bit wired
to me. But maybe there are good reasons for this so I didn't modified
that behaviour.

But I added a check for the rsync return code. Before my PR it was
ignored and had no effect when rsync failed.
My testing class directly derive from `unittest.TestCase` and do not use
one of the classes in `generic.py`. The latter isn't an elegant solution
but this wasn't the reason why I avoided there use. I needed to
introduce blanks in the path names to trigger the "rsync problem I
need".

I collected all the code fragments from the `setUp()` etc methods in
`generic.py` and build my own _helper_ methods. I copied and refactored
that code into private methods of my testing class. I'm sure we can
improve that in the future and find a more _generic_ solution to reuse
that fragments for other unittests also. But it is IMHO a good start and
also helped me to improve my understanding of BIT internals.
Sorry I tried to resist but I did some code formating. I also added some
docstrings while trying to understand some details.

Related: #1247