Test backing up permissions in a "SSH local" profile.
Introducing class `test_snapshots.py::TestSshPermissions` testing
`snapshots.py::Snapshot.backupPermissions()` in a "SSH local"
environment.

The case was untested before. Especially it is needed to support the fix
of #1247 (rsync incompatibility).

I slightly improved the method `Snapshots.backupPermissions()` because
before it was silent and didn't care about if the used `rsync` call
succeeded or not. No the return code of `rsync` is returned and can be
used in a unittest.

The "helper" functions in `test_snapshots.py` are refactored and
improved a bit. Now they are used by this new `TestSshPermissions` and
the before introduced `TestSshRemoveSnapshot` tests.