Fix: Incompatibility with rsync 3.2.4 or newer
The core problem is that BIT do call rsync in a way that is not compatible
with newer versions of rsync. The version 3.2.4 of rsync activated a new
method of argument protection. This do fix the rsync calls and do all
necessary modifications.

Important to know is that BIT still work with old rsync versions, too.

Fix #1247