Fix: Incompatibility with rsync 3.2.4++ due to changed arg protection (#1351)
The core problem is that BIT does call rsync in a way that is not compatible
with newer versions of rsync. The version 3.2.4 of rsync activated a new
method of argument protection. This commit fixes the rsync calls and does all
necessary modifications.

Important to know is that BIT still works with old rsync versions, too.

Fix #1247