fix!: SYSLOG_IDENTIFIER is "backintime"
Always use "backintime" as identifier for syslog entries.

Previously syslog (or a component behind it) silently refused the identifier because of a whitespace character in it. The name and ID of the snapshot profile is now part of the message string instead of the identifier. Even the GUI component do use "backintime" as identifier.

BREAKING CHANGE: This does modify the syntax of syslog entries.

Fix #1483
Close #1451