fix: Global flock fallback to single-user mode if insufficient permissions (#1743, #1751)
The global flock context manager (introduced in #1679) now fall back to single-user mode if permissions are insufficient to create an flock file accessible by other users and root. The multi-user mode will be activate later and automatically when BIT is started as root for the first time. It will create an flock file accessible by all users.

That behavior is relevant for Arch-based systems but not relevant on Debian-based systems because regular users have enough permissions to /run/lock.

Additionally fix a bug (introduced with #1679) ignoring the config value of `global.use_flock`.

Additionally Sphinx config is modified and improved..

Fix #1743 
Fix #1751