feat: Check and warn if include list entries don't exist in backup source
In case files and directories included in the backup profile do not
exist in the backup source the user get a warning. The warning is
published in the the log files and in the systray (if available). The
main window do not warn because of technical obstacles that will be
fixed later.

Close #1586

Thank you very much to @rafaelhdr for this big contribution.
---------

Co-authored-by: Rafael Hurpia da Rocha <1108973+rafaelhdr@users.noreply.github.com>
Co-authored-by: buhtz <c.buhtz@posteo.jp>
Co-authored-by: Rafael Hurpia da Rocha <639275-rafaelhdr@users.noreply.gitlab.com>