fix: Prevent name conflict in config method
The config module imports `password`. Previously, the `setPassword`
method used the conflicting argument name `password`, which shadowed the
name of the imported module. It also attemptd to _use_ the shadowed
module name if `self.pw` was None, which means that sometimes the method
would work, and sometimes it wouldn't.

I've resolved this by renaming the argument name to `password_value`.