Fix SSH mode broken for new users:
Issue: In commit Commit 1ecedfa, a change was made to run ssh-keygen with 'SSH_ASKPASS_REQUIRE': 'prefer'.
However, this used up the password placed in the FIFO that was put there for the use in the following invocation of ssh-add.
This caused the ssh-add to not have a waiting password in the FIFO, resulting in a failure.

Fix:  Add a new filling of the FIFO after the ssh-keygen check.