fix: Unlocking SSHkey with passphrase on new created SSH profiles (#2163)
A change (1ecedfa) was made to run ssh-keygen with 'SSH_ASKPASS_REQUIRE': 'prefer'.
However, this used up the password placed in the FIFO that was put there for the use in the following invocation of ssh-add.
This caused the ssh-add to not have a waiting password in the FIFO, resulting in a failure.

This commit fix this problem with filling of again the FIFO after the ssh-keygen check.

Bug introduced with 1ecedfa and released in v1.5.4.
Thank you to @davidfjoh for analyzing the bug and providing this fix.
Keeping #2164 open until the fix arrived in hotfix release v1.5.5.
---------
Co-authored-by: Christian Buhtz <c.buhtz@posteo.jp>