feat(gui): Systray icon can be forced to Dark or Light
User is able, to force a dark or white fill color for symbolic icon used in the systray.
By default the fill color is determined based on the theme used by the desktop environment which can lead to false combination of colors under several circumstances.

The setting is located in the main menu bar as submenu "Systray icon" in the "Back In Time" menu. It is planed to move it into an application settings dialog later which do not exist yet.

PR is a follow up of PR #2293 (899db58) and PR #2278 (31970ae).